+++ src/ngi_projecter/__init__.py
@@ -1,2 +1,0 @@
-from ngi_projecter.projecter import *
-from ngi_projecter.projections import *